Atualização no package.json e package-lock.json - Adição via npm da biblioteca framer-motion
Atualização do componente Botão
- Adição do loading (animação) ao recuperar os dados da API do Twitter
- Adição de mensagem ao inserir um texto a mais no campo busca
- Adição de mensagem ao tentar realizar a busca sem nenhum texto
- Adição do botão "ENTER" como um dos processos de pesquisa do input search
