+++ hashtagfinder/package.json
@@ -9,6 +9,7 @@
     "airtable": "^0.11.4",
     "axios": "^0.27.2",
     "formik": "^2.2.9",
+    "framer-motion": "^6.3.16",
     "netlify-cli": "^10.7.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
